Overhaul to clustering outputs
Clustering now writes clusters to a single .npz.ag file instead of PDB files for individual
residues. This is primarily to accommodate applying and saving custom data attributes to
individual motif residues in each cluster. This made a lot of code that kept track of motif
residue source redundant. Residue source and other stuff is applied to individual motif
residues at this step.