Complete overhaul to how fuzzballs are assembled
PyRosetta is now used to score and assemble fuzzballs. This is implemented in the Generate_Fuzzball_using_PyRosetta
class. Clusters are now transformed for each conformer and ligand-residue interactions are scored. These scores
are used to populate fuzzballs with the best X residues based on REU, versus previously where "best" clusters were
used to populate the fuzzball only for most to be thrown out at the FeatureReporter/Gurobi steps.

A framework for adding user-defined residues to the fuzzball has been implemented. However, the ability to use this
for generating new fuzzballs for iterations has not yet been implemented.