Refactored how clustering is performed due to limitations in agglomerative heirarchical clustering
This clustering method is O(n^2) in memory, which means it will blow up when working with fragments that are highly represented in the PDB. To get around this, I first cluster using residue categorical characteristics len 14 with a hamming distance cutoff of 2. These clusters should contain residues all making the same type of interactions with the fragment. I then cluster using the contact vector by cosine distance. This should produce the same result as the previous method, but more clearly defined.