Alignments can optionally take fragment connectivity in the source molecule into consideration when mapping the fragment onto substructure-containing molecules found in the PDB.
This is to ensure that mappings only occur with substructures that possess similar local chemistry as the source molecule. The substructure mapping algo in RDKit will map substructures even if substructure atoms possess different substituents/valences.

I also updated Alignments to keep track of rejected PDBs per fragment in a pickle instead of the global rejected text file from before. A reject for one fragment may not mean the entire PDB is bad.